refactor: split coord map code into modules
