+++ packages_rs/nextclade/src/run/nextalign_run_one.rs
@@ -1,7 +1,7 @@
 use crate::align::align::align_nuc;
 use crate::align::insertions_strip::{get_aa_insertions, insertions_strip};
 use crate::align::params::AlignPairwiseParams;
-use crate::coord::coord_map::CoordMapGlobal;
+use crate::coord::coord_map_global::CoordMapGlobal;
 use crate::io::gene_map::GeneMap;
 use crate::io::nuc::Nuc;
 use crate::translate::translate_genes::{translate_genes, Translation};
